# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Ecotrack API
  version: 1.0.0
  description: Documentazione REST API per Ecotrack. Generata automaticamente da swagger-jsdoc.
  contact:
    name: Team Ecotrack
servers:
  - url: http://localhost:3000
    description: Server di sviluppo locale
paths:
  /session/login:
    post:
      summary: Log in a user via an auth provider
      tags:
        - Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - provider_user_id
                - provider_data
                - provider_type_name
              properties:
                provider_user_id:
                  type: string
                  example: john@example.com
                provider_data:
                  oneOf:
                    - $ref: '#/components/schemas/LocalProviderData'
                  description: |
                    Provider-specific auth payload. For "local": { password: string }.
                provider_type_name:
                  type: string
                  example: local
            examples:
              local:
                summary: Login con email e password
                value:
                  provider_user_id: john@example.com
                  provider_data:
                    password: s3cr3tP@ssword
                  provider_type_name: local
      responses:
        '200':
          description: Login successful. Sets a session cookie and returns the token.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session=eyJ...; HttpOnly; SameSite=Lax; Path=/
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: All fields are required
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error — session could not be created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/logout:
    post:
      summary: Log out the current user
      tags:
        - Session
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful. Clears the session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Logged out successfully
  /session/validate:
    get:
      summary: Validate the current session cookie
      tags:
        - Session
      security:
        - cookieAuth: []
      responses:
        '200':
          description: |
            Always returns 200. Check the `valid` field to determine whether the session is active.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      valid:
                        type: boolean
                        example: true
                      user:
                        $ref: '#/components/schemas/User'
                  - type: object
                    properties:
                      valid:
                        type: boolean
                        example: false
  /session/refresh:
    post:
      summary: Refresh the current session token
      tags:
        - Session
      security:
        - cookieAuth: []
      responses:
        '200':
          description: >
            Always returns 200. On success a new session cookie is set. Check the `message` field for details on failure
            cases.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session=eyJ...; HttpOnly; SameSite=Lax; Path=/
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              examples:
                refreshed:
                  value:
                    message: Session refreshed
                noSession:
                  value:
                    message: No session found
                invalidSession:
                  value:
                    message: Invalid or expired session
  /session/me:
    get:
      summary: Get the currently authenticated user
      tags:
        - Session
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Returns the authenticated user's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Missing or invalid session cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                noSession:
                  value:
                    message: No session found
                invalidSession:
                  value:
                    message: Invalid or expired session
        '404':
          description: Authenticated user ID not found in the database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: User not found
  /user/register:
    post:
      summary: Register a new user with local email/password auth
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: John Doe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: s3cr3tP@ssword
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: User created successfully
        '400':
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingFields:
                  value:
                    message: All fields are required
                invalidEmail:
                  value:
                    message: Invalid email format
                passwordTooShort:
                  value:
                    message: Password must be at least 8 characters
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Email already in use
        '500':
          description: Database error while creating the user or auth provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                userCreationFailed:
                  value:
                    message: Could not create user
                authProviderFailed:
                  value:
                    message: Could not create auth provider
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: john@example.com
    AuthToken:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation successful
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: An error occurred
    LocalProviderData:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          format: password
          example: s3cr3tP@ssword
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
tags: []